fix swipe left show dictionary
